大数据导论项目 - 副本/requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.8.4
scikit-learn==1.4.2
openpyxl==3.1.5
tabulate==0.9.0
langgraph==0.2.35
openai==1.40.0

